# Or Selector benchmark and removal of repeated virtual calls