I had similar changes like in https://github.com/AvaloniaUI/Avalonia/pull/10478 but I didn't have time to wrap it up before leaving for holidays. I've started from writing a benchmark case so we can avoid this issue in the future. More selectors could be covered later too.

## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Removes repeated virtual calls to `Count`.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
There are many repeated virtual calls to `Count` that can be easily avoided.
Also it might be the case that a few changes to `StyledElement` might have the same problem, in one case in `ApplyStyles` it might be an actual pessimization to use a for loop since `Styles` have a struct enumerator that does not allocate anything.

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
Minor perf increase, I didn't bother to measure. But I have benchmarks for or selector with optimization to iteration (as introduced in https://github.com/AvaloniaUI/Avalonia/pull/10478):

Before
|                Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|---------------------- |---------:|---------:|---------:|-------:|----------:|
|  OrSelector_One_Match | 50.95 ns | 0.223 ns | 0.198 ns | 0.0076 |      32 B |
| OrSelector_Five_Match | 52.05 ns | 0.265 ns | 0.248 ns | 0.0076 |      32 B |

After
|                Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|---------------------- |---------:|---------:|---------:|-------:|----------:|
|  OrSelector_One_Match | 39.70 ns | 0.315 ns | 0.263 ns |      - |         - |
| OrSelector_Five_Match | 39.79 ns | 0.329 ns | 0.275 ns |      - |         - |

So we are saving 32B per match while also getting ~20% perf speedup. So yeah an analyzer would help a lot since this little issue causes trouble all over the place.

Also added extra run configs for benchmark project so we can easily control it without manually changing arguments.

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->


## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Documentation with user documentation

## Breaking changes
<!--- List any breaking changes here. -->

## Obsoletions / Deprecations
<!--- Obsolete and Deprecated attributes on APIs MUST only be included when discussed with Core team. @grokys, @kekekeks & @danwalmsley -->

## Fixed issues
<!--- If the pull request fixes issue(s) list them like this: 
Fixes #123
Fixes #456
-->