@MrJul 2023-03-06
Thanks for providing a benchmark! (I'm the author of #10478)
The virtual call reductions are nice, I'll keep that in mind for future changes :)


> Also it might be the case that a few changes to `StyledElement` might have the same problem, in one case in `ApplyStyles` it might be an actual pessimization to use a for loop since `Styles` have a struct enumerator that does not allocate anything.

I went with `for` for consistency here: indeed the `Styles` enumerator won't be boxed. That said, `Styles.Count` isn't virtual either and cheap enough to be inlined so there's no heavy cost to pay here. (See [this example](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8lAbhvqfIDoBhCADYCYYDAEsIAO1zcA8sABWIjAFkIAExgD21DgGYmpBrwYBvGg0tMDxFAwAyY3Bn4BXSRgAUj5wB5mdAB8DAJOGACUZhZWMQBm0AyeAG7YUAxiDAC8DHSs6Qy+IWF8EO4YeQDUFWLh0TH1XACcnqHOANpiALrhOjEAvnUMg8Q2dj4upR72EGDYAt5h/uRBRc6R5tQNlilpkGVZqxglZb1b8WnJqfnZufmFex6V1bWbWzHMza0YHd2nlgOvSzDUYOMIAMWgMGwYAAFgs/AFgl91oM4pDoTDEjt0hgYPh0pJDi83u9yJ8wj1BgCATQgA=) where you can see that `Count` has been inlined to a `[rsi+0x10]` access.)

The other cases could definitely cache the virtual call count.